调整类型

+++ d3-base/d3-base-core/src/main/java/com/github/d3/base/entity/Action.java
@@ -38,7 +38,7 @@ public class Action extends MpBaseEntity {
      * 动作值
      */
     @TableField("val")
-    private Long val;
+    private Integer val;
 
     /**
      * 模块ID
